LayoutTests/fast/media/mq-any-hover-cssom-expected.txt

Test the CSSOM serialization of the any-hover media feature.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: none)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: none)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: hover)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(any-hover: ON-DEMAND)"
PASS document.styleSheets[1].cssRules[0].media.mediaText is "(Any-Hover: On-demand)"
PASS successfullyParsed is true

TEST COMPLETE
